refactor: todo items screen
